requirements.txt
aiofiles
aiohttp
apscheduler
asyncio
bard
beautifulsoup4
cloudscraper
daxxhub
deepai
dnspython
emojis
Faker
ffmpeg-python
future==0.18.3
git+https://github.com/AsmSafone/SafoneAPI
gitpython
gpytranslate
gtts
hachoir
heroku3
httpx==0.27.2
img2pdf
lexica-api==1.5.1
motor
ntgcalls==1.1.2
numpy
openai
opencv-python
pickledb
pillow==9.5.0
psutil
pycountry
pycryptodome
pydub
pyfiglet
pykeyboard
pyshorteners
pyrofork
python-dotenv
python-telegram-bot==21.1
python-whois
pytz
pyshorteners
py-tgcalls==0.9.7
pyyaml
qrcode
requests
search_engine_parser
speedtest-cli
spotipy
SpeechRecognition
telegraph
telethon
tgcrypto
unidecode
uvloop
wget
youtube-search
youtube-search-python
yt-dlp>=2025.6.25
https://github.com/KurimuzonAkuma/pyrogram/archive/v2.1.23.zip
